Add a heap more things
It's kinda crazy just how heavy the runTurn method is, but I don't think there's really a sensible way to break it up further. I added a heap of things around it, but the remaining methods will have to wait. Also bugfixed a few things, and changed the bomb locations to be stamped directly onto the board rather than being on a separate overlay because that was a little weird.