Money Cards rebalance
The values themselves have been slashed in half except for the starting minimum bet and required bet multiple, which are now $2,000 instead of the expected $2,500 (I've also bounced $1,000 as an idea). A bonus score double has been added for a perfect game, though, so the jackpot is unchanged from $2,880,000.

Also among the changes is that "all" has been added as an alias for "all-in" and "all in".